Huge refac
I am trying a new model. This should be fairly close to C++ spirit,
with almost no pointers and play nice with cuda.
Basically I am using template specilaization for the api.
The idea is the following. Whether we are querying the
scene primitive or the scene group, the type enums
can be used in conjunction with the templates.
Some thing like
hit<SceneGroup>(sceneGroup){
if(gtype == SOLID){
  return hit<SOLID>(sceneGroup)
  }else{
  return hit<CONSTANT_MEDIUM>(sceneGroup)
  }
}
this would call hit<ScenePrimitive>(ScenePrim, etc) on all
primitives which in turn would be something like
hit<ScenePrimitive>(ScenePrim){
  if (scenePrim.htype == XY_RECT){
    return hit<XY_RECT>(ScenePrim)
  } else {
    return hit<SPHERE>(scenePrim)
  }
}
or for scattering
scatter<HitRecord>(HitRecord rec, dist, ray){
  if (rec.material_params.mtype == LAMBERT){
    return scatter<LAMBERT>(rec, dist, ray)
  }
}