Another huge refactoring compile errors mostly gone
Now most of the compile errors are gone.
We should concentrate how we would handle the scene.
The main problem is loading pictures.
Textures needs to access Scene in order to
be used by materials, if we want to use images,
and not copy them etc.

The way I thought about it is the following:

The hit is done at scene level.
If there is any hit, we store the objects index,
inside the hit record.
Using that index we check if the object is an
emissive material if it is not emissive we
give 0.
If it is emissive we apply emissive. We initialize
the scene material for the given material param.
if its associated texture parameter is of IMAGE
type, we get the necessary color value from
scene's image unsigned char array, with given
info in material params.

For scattering we do the exact same thing.
Initialize the scene material for the given
material param.
If its associated texture is of IMAGE type,
we get its albedo from unsigned char array,
if not we obey the scene material procedure.
So the scene material procedure covers everything
except the image type.