Huge refactoring under way.
The hit method and other methods of the sceneobjs need
a lot of refactoring.

The new idea is instead of recreating the group object
for collision check, we call the hit with group index.
something like hit_group(group_index, ray etc)
then we check for each primitive.

For constant medium and the like who has specific
hit, scattering functions for group.
We can check the hit record who has a field for
group_scattering. If that field is true,
we can call group scattering, if not
primitive scattering.